refactor: created RootProvider component
