+++ src/components/ContextHOCs/UsersTasksWithContext.jsx
@@ -1,4 +1,4 @@
-import { UserContext } from '../../App/userContext';
+import { UserContext } from '../RootProvider/userContext';
 import UsersTasks from '../../pages/UsersTasks';
 
 export default function UsersTasksWithContext(props) {
