+++ src/components/RootProvider/themeContext.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { THEMECONTEXT } from '../utilities/context-helpers';
+import { THEMECONTEXT } from '../../utilities/context-helpers';
 
 export const ThemeContext = React.createContext();
