+++ src/components/RootProvider/userContext.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import noop from '../shared/noop';
-import { USERCONTEXT } from '../utilities/context-helpers';
+import noop from '../../shared/noop';
+import { USERCONTEXT } from '../../utilities/context-helpers';
 
 export const UserContext = React.createContext({
   isLogged: false,
